Add to order post, put, delete and dto

+++ src/products/products.controller.ts
@@ -1,6 +1,7 @@
-import { Controller, Delete, Get, Param, Post, Body } from '@nestjs/common';
+import { Controller, Delete, Get, Param, Post, Put,    Body,  ParseUUIDPipe, NotFoundException } from '@nestjs/common';
 import { ProductsService } from './products.service';
-import { CreateProductDTO } from './create-product.dto';
+import { CreateProductDTO } from './dtos/create-product.dto';
+import { UpdateProductDTO } from './dtos/update-product.dto';
 
 @Controller('products')
 export class ProductsController { 
@@ -12,8 +13,10 @@ export class ProductsController {
   }
 
   @Get('/:id')
-  public getById(@Param('id') id: string) {
-    return this.productsService.getById(id);
+  getById(@Param('id', new ParseUUIDPipe()) id: string) {
+    const prod = this.productsService.getById(id);
+    if (!prod) throw new NotFoundException('Product not found');
+    return prod;
   }
 
   @Post('/')
@@ -21,9 +24,24 @@ export class ProductsController {
     return this.productsService.create(productData);
   }
 
+  @Put('/:id')
+  update(
+    @Param('id', new ParseUUIDPipe()) id: string,
+    @Body() productData: UpdateProductDTO,
+  ) {
+    if (!this.productsService.getById(id))
+      throw new NotFoundException('Product not found');
+
+    this.productsService.updateById(id, productData);
+    return { success: true };
+  }
+
   @Delete('/:id')
-  public deleteProduct(@Param('id') id: string) {
-    this.productsService.deleteProduct(id);
-    return { success: true }
+  deleteProduct(@Param('id', new ParseUUIDPipe()) id: string) {
+  if (!this.productsService.getById(id))
+    throw new NotFoundException('Product not found');
+
+  this.productsService.deleteProduct(id);
+  return { success: true };
   }
 }
